requirements.txt
python-socketio==5.6.0
uvicorn==0.18.2
tk==0.1.0
speedtest-cli==2.1.3
user_agent==0.1.10
requests==2.31.0